refactor(index): move helper functions to separate files
closes #824
